Add Dockerfile and requirements.txt
It is the first commit. The docker file contains the commands to
create a development environment for odoo, and the requirements.txt has
the python libraries which is the exact requirements.txt in the odoo
source code.
The Dockerfile content has been taken from https://github.com/odoo/docker.git
and modified to it's current content to satisfy the development
requiremnts.